atom.xml: add post image to top for feedwind
The hacked up atom feed actually worked out of the gate with feedwind,
which is great. But it appears they grab whatever the first image they
find is and use it as a thumbnail. So, i've added the post image to the
top to make sure it is teh first image and is used as the picture, if
anything