Major abuse to the original import script to serve this new purpose - #13
Made the author handling logic in the import extremely lax in order to
get it to actually run, since many author changes have been made since
the time of the author data export. This would be terrible except the
case we are now running the script is an already populated database, and
we're just trying to fix a very small handful of missed posts do to the
missing body malformed data scenario. So, none of the existing
validation, which was very, very important at the begining, doesn't
really matter now since the posts which it would be creating already
exist